+++ apps/frontend-manage/src/lib/utils/filters.ts
@@ -1,40 +1,24 @@
-/* eslint-disable no-unused-vars */
-// TODO: typing
-
 import dayjs from 'dayjs'
 import _every from 'lodash/every'
 // import Fuse from 'fuse.js'
 import { Element } from '@klicker-uzh/graphql/dist/ops'
+import {
+  QuestionPoolFilters,
+  QuestionPoolSortType,
+  SortyByType,
+} from '@lib/hooks/useSortingAndFiltering'
 import * as JsSearch from 'js-search'
-
-const indices = {}
-
-/* const INDEX_CONFIGS = {
-  questions: {
-    location: 0,
-    distance: 100,
-    threshold: 0.5,
-    maxPatternLength: 32,
-    minMatchCharLength: 1,
-    tokenize: true,
-    matchAllTokens: true,
-    includeScore: true,
-    includeMatches: true,
-    // TODO: add description to search keys?
-    keys: ['title', 'tags.name', 'type'],
-  },
-} */
 
 export function buildIndex(
   name: string,
-  items: any[],
-  searchIndices: any[]
-): any[] {
+  items: Element[],
+  searchIndices: string[]
+): JsSearch.Search {
   // build a new js-search index
   const search = new JsSearch.Search('id')
 
   // use the TF-IDF strategy
-  search.searchIndex = new JsSearch.TfIdfSearchIndex()
+  search.searchIndex = new JsSearch.TfIdfSearchIndex('id')
 
   // look for all substrings, not only prefixed
   search.indexStrategy = new JsSearch.AllSubstringsIndexStrategy()
@@ -52,22 +36,20 @@ export function buildIndex(
   // build the index based on the items
   search.addDocuments(items)
 
-  // store the index and return it (singleton)
-  indices[name] = search
   return search
 }
 
 // TODO: optimize for one pass instead of stacked passes
 export function filterQuestions(
-  questions: any[],
-  filters: any,
-  index: any
-): any[] {
+  questions: Element[],
+  filters: QuestionPoolFilters,
+  index: JsSearch.Search | null
+): Element[] {
   let results = [...questions]
 
   // if a title (query) was given, search the index with it
   if (index && filters.name) {
-    results = index.search(filters.name)
+    results = index.search(filters.name) as Element[]
   }
 
   // only reduce the shown questions to the non-archived ones, if the archive filter is not active
@@ -82,14 +64,14 @@ export function filterQuestions(
   // if a sample solution filter was selected, only show questions with a sample solution
   if (filters.sampleSolution) {
     results = results.filter(
-      ({ hasSampleSolution }): boolean => hasSampleSolution === true
+      ({ options }) => options.hasSampleSolution === true
     )
   }
 
   // if an answer feedback filter was selected, only show questions with answer feedbacks
   if (filters.answerFeedbacks) {
     results = results.filter(
-      ({ hasAnswerFeedbacks }): boolean => hasAnswerFeedbacks === true
+      ({ options }) => options.hasAnswerFeedbacks === true
     )
   }
 
@@ -104,8 +86,10 @@ export function filterQuestions(
       // compare the tags selected and check whether the question fulfills all of them
       if (
         filters.tags &&
-        !_every(filters.tags, (tag): boolean =>
-          tags.map((t): string => t.name).includes(tag)
+        !_every(
+          filters.tags,
+          (tag): boolean =>
+            tags?.map((t): string => t.name).includes(tag) ?? false
         )
       ) {
         return false
@@ -118,69 +102,62 @@ export function filterQuestions(
   return results
 }
 
-// both used for sessions and users
-export function filterByTitle(objects: any[], filters: any, index: any): any[] {
-  let results = objects
-
-  if (filters.title) {
-    results = index.search(filters.title)
-  }
-
-  return results
-}
-
-export function subtractDates(date1: any, date2: any): any {
-  return date1 - date2
-}
-
-export function sortQuestions(questions: any[], sort: any): any[] {
+export function sortQuestions(
+  questions: Element[],
+  sort: QuestionPoolSortType
+): Element[] {
   const factor = sort.asc ? 1 : -1
 
-  if (sort.by === 'TITLE') {
+  if (sort.by === SortyByType.TITLE) {
     return questions.sort(
       (a, b): number => factor * a.name.localeCompare(b.name)
     )
   }
 
-  if (sort.by === 'TYPE') {
+  if (sort.by === SortyByType.TYPE) {
     return questions.sort(
       (a, b): number => factor * a.type.localeCompare(b.type)
     )
   }
 
-  if (sort.by === 'CREATED') {
+  if (sort.by === SortyByType.CREATED) {
     return questions.sort(
-      (a, b): number =>
-        factor * subtractDates(dayjs(a.createdAt), dayjs(b.createdAt))
+      (a, b): number => factor * dayjs(a.createdAt).diff(dayjs(b.createdAt))
     )
   }
 
-  if (sort.by === 'USED') {
-    return questions.sort((a, b): number => {
-      if (a.instances.length === 0 || b.instances.length === 0) {
-        if (a.instances.length === 0 && b.instances.length === 0) {
-          return 0
-        }
+  // TODO: if desired, fetch instances / number of instances as well and re-introduce this option
+  // if (sort.by === SortyByType.USED) {
+  //   return questions.sort((a, b): number => {
+  //     if (a.instances.length === 0 || b.instances.length === 0) {
+  //       if (a.instances.length === 0 && b.instances.length === 0) {
+  //         return 0
+  //       }
 
-        return factor * 1
-      }
+  //       return factor * 1
+  //     }
 
-      // compare the dates of the latest created instances
-      // this allows us to sort by "last usage"
-      return (
-        factor *
-        subtractDates(
-          dayjs(a.instances[a.instances.length - 1].createdAt),
-          dayjs(b.instances[b.instances.length - 1].createdAt)
-        )
-      )
-    })
-  }
+  //     // compare the dates of the latest created instances
+  //     // this allows us to sort by "last usage"
+  //     return (
+  //       factor *
+  //       subtractDates(
+  //         dayjs(a.instances[a.instances.length - 1].createdAt),
+  //         dayjs(b.instances[b.instances.length - 1].createdAt)
+  //       )
+  //     )
+  //   })
+  // }
 
   return questions
 }
 
-export function processItems(items, filters, sort, index): Element[] {
+export function processItems(
+  items: Element[],
+  filters: QuestionPoolFilters,
+  sort: QuestionPoolSortType,
+  index: JsSearch.Search | null
+): Element[] {
   let processed = items
 
   if (filters) {