+++ apps/frontend-pwa/src/pages/course/[courseId]/quiz/[id].tsx
@@ -40,7 +40,7 @@ function PracticeQuizPage({ courseId, id }: Props) {
       <Layout>
         <UserNotification
           type="error"
-          message={t('pwa.learningElement.notFound')}
+          message={t('pwa.practiceQuiz.notFound')}
         />
       </Layout>
     )
@@ -67,7 +67,7 @@ function PracticeQuizPage({ courseId, id }: Props) {
         showResetLocalStorage
       />
       <Footer
-        browserLink={`${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/element/${id}`}
+        browserLink={`${process.env.NEXT_PUBLIC_PWA_URL}/course/${courseId}/quiz/${id}`}
       />
     </Layout>
   )