+++ packages/graphql/src/services/participants.ts
@@ -1,6 +1,6 @@
 import {
-  LearningElementStatus,
   MicroSessionStatus,
+  PublicationStatus,
   SessionStatus,
 } from '@klicker-uzh/prisma'
 import bcrypt from 'bcryptjs'
@@ -401,40 +401,6 @@ export async function getParticipation(
 //   }
 // }
 
-// TODO: remove this after migration to practice quiz
-interface BookmarkQuestionArgs {
-  stackId: number
-  courseId: string
-  bookmarked: boolean
-}
-
-// TODO: remove this function after migration to practice quiz
-export async function bookmarkQuestion(
-  { stackId, courseId, bookmarked }: BookmarkQuestionArgs,
-  ctx: ContextWithUser
-) {
-  const participation = await ctx.prisma.participation.update({
-    where: {
-      courseId_participantId: {
-        courseId,
-        participantId: ctx.user.sub,
-      },
-    },
-    data: {
-      bookmarkedStacks: {
-        [bookmarked ? 'connect' : 'disconnect']: {
-          id: stackId,
-        },
-      },
-    },
-    include: {
-      bookmarkedStacks: true,
-    },
-  })
-
-  return participation.bookmarkedStacks
-}
-
 interface BookmarkElementStackArgs {
   stackId: number
   courseId: string
@@ -465,39 +431,6 @@ export async function bookmarkElementStack(
   })
 
   return participation.bookmarkedElementStacks.map((stack) => stack.id)
-}
-
-// TODO: remove after migration
-interface GetBookmarkedQuestionsArgs {
-  courseId: string
-}
-
-// TODO: remove after migration
-export async function getBookmarkedQuestions(
-  { courseId }: GetBookmarkedQuestionsArgs,
-  ctx: ContextWithUser
-) {
-  const participation = await ctx.prisma.participation.findUnique({
-    where: {
-      courseId_participantId: {
-        courseId,
-        participantId: ctx.user.sub,
-      },
-    },
-    include: {
-      bookmarkedStacks: {
-        include: {
-          elements: {
-            include: {
-              questionInstance: true,
-            },
-          },
-        },
-      },
-    },
-  })
-
-  return participation?.bookmarkedStacks ?? []
 }
 
 interface GetBookmarkedElementStacksArgs {
@@ -541,19 +474,6 @@ export async function flagQuestion(
       id: args.questionInstanceId,
     },
     include: {
-      stackElement: {
-        include: {
-          stack: {
-            include: {
-              learningElement: {
-                include: {
-                  course: true,
-                },
-              },
-            },
-          },
-        },
-      },
       microSession: {
         include: {
           course: true,
@@ -562,12 +482,7 @@ export async function flagQuestion(
     },
   })
 
-  if (
-    !questionInstance?.stackElement?.stack?.learningElement?.course
-      ?.notificationEmail &&
-    !questionInstance?.microSession?.course?.notificationEmail
-  )
-    return null
+  if (!questionInstance?.microSession?.course?.notificationEmail) return null
 
   await fetch(process.env.NOTIFICATION_URL as string, {
     method: 'POST',
@@ -576,23 +491,15 @@ export async function flagQuestion(
       'Content-Type': 'application/json',
     },
     body: JSON.stringify({
-      elementType: questionInstance?.stackElement?.stack.learningElement
-        ? 'Practice Quiz'
-        : 'Microlearning',
-      elementId:
-        questionInstance?.stackElement?.stack?.learningElement?.id ||
-        questionInstance?.microSession?.id,
-      elementName:
-        questionInstance?.stackElement?.stack?.learningElement?.name ||
-        questionInstance?.microSession?.name,
+      elementType: 'Microlearning',
+      elementId: questionInstance?.microSession?.id,
+      elementName: questionInstance?.microSession?.name,
       questionId: questionInstance.questionId,
       questionName: questionInstance.questionData.name,
       content: args.content,
       participantId: ctx.user?.sub,
       secret: process.env.NOTIFICATION_SECRET,
       notificationEmail:
-        questionInstance.stackElement?.stack?.learningElement?.course
-          ?.notificationEmail ||
         questionInstance.microSession?.course?.notificationEmail,
     }),
   })
@@ -758,14 +665,9 @@ export async function getPracticeQuizList(ctx: ContextWithUser) {
     include: {
       course: {
         include: {
-          learningElements: {
-            where: {
-              status: LearningElementStatus.PUBLISHED,
-            },
-          },
           practiceQuizzes: {
             where: {
-              status: LearningElementStatus.PUBLISHED,
+              status: PublicationStatus.PUBLISHED,
             },
           },
         },
@@ -778,11 +680,7 @@ export async function getPracticeQuizList(ctx: ContextWithUser) {
   const courses = participations
     .map((p) => p.course)
     .sort((a, b) => (a.endDate > b.endDate ? -1 : 1))
-    .filter(
-      (course) =>
-        course.practiceQuizzes.length !== 0 ||
-        course.learningElements.length !== 0
-    )
+    .filter((course) => course.practiceQuizzes.length !== 0)
 
   return courses
 }